requirements.txt
gensim==3.8.2
streamlit==0.58.0
libvoikko==4.3
s3fs==0.4.2
seaborn==0.10.1
matplotlib==3.2.1
pytest==5.4.1
botocore==1.15.47
fastparquet==0.3.3
pandas==1.0.3
boto3==1.12.47
more_itertools==8.2.0
Markdown==3.2.1
requests==2.23.0
psycopg2==2.8.4
bokeh==2.0.2
voikko==0.5
plotly==4.6.0
nltk==3.5
numpy==1.18.3
umap_learn==0.4.1
pyLDAvis==2.1.2
PyYAML==5.3.1
scikit_learn==0.22.2.post1